fix paths, rename multiple of to paginated

+++ openapi/model_quote.go
@@ -13,9 +13,9 @@ package openapi
 type Quote struct {
 	AdjustedPreviousClose string `json:"adjusted_previous_close,omitempty"`
 	AskPrice string `json:"ask_price,omitempty"`
-	AskSize float32 `json:"ask_size,omitempty"`
+	AskSize int32 `json:"ask_size,omitempty"`
 	BidPrice string `json:"bid_price,omitempty"`
-	BidSize float32 `json:"bid_size,omitempty"`
+	BidSize int32 `json:"bid_size,omitempty"`
 	LastExtendedHoursTradePrice string `json:"last_extended_hours_trade_price,omitempty"`
 	LastTradePrice string `json:"last_trade_price,omitempty"`
 	PreviousClose string `json:"previous_close,omitempty"`
